+++ go/cwc/io/morse.go
@@ -1,0 +1,14 @@
+package io
+
+import "time"
+
+import "../../bitoip"
+
+// Absolute time morse event
+type TimedBitEvent struct {
+	startTime time.Time
+	bitEvent  bitoip.BitEvent
+}
+
+// slice of events
+type MorseEvents []TimedBitEvent
