+++ go/cwc/lang/encoder.go
@@ -15,7 +15,11 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-package io
+package lang
 
-// Morse text decoder
-// Uses the generic decoder to output from a morse stream to an io.Writer
+import "context"
+
+// Morse text encoder
+// Given input as an io.Reader, turn into a Morse stream
+
+func StartEncoder(ctx context.Context, textIn Reader)
